Add a commerce cart form for use by `registration` order types
Includes some participant info and excludes any quantity settings.

Note that commerce cart form views are normally use the orders table as the base, while this view one uses the registration table as the base with relationships with the order_items and orders tables. This allows the display of participants and other registration-related info, but it also means that it can only be used with product variations that will be added to the cart via erf_commerce integration. This is why the previous commits set some very specific order item and order types.