Add a checkout order summary view for use by `registration` order types
Includes some participant info and excludes any quantity settings.

Note that commerce checkout order summary views are normally use the orders table as the base, while this one uses the registration table as the base and a relationship with the order_items table. This allows the display of participants and other registration-related info, but it also means that it can only be used with product variations that will be added to the cart via erf_commerce integration. This is why the previous commits set some very specific order item and order types.

Note also that the view_mode for the purchased entity field was set to `cart` manually in the config file because the views UI shows the wrong options (see https://www.drupal.org/project/commerce/issues/2638958).