Fix missing items in order view
Commerce 2.24 included a fix that changed the arguments (Contextual filter) for the order item view. We deal with most of the effects of that change in commit e7cbceb, but missed the 'Allow multiple values' setting in the contextual filter.

This is probably because, for some puzzling reason, that setting has the machine name 'break_phrase'.

Without this setting enabled, only the first item in the order will be displayed on the order detail and completion pages.