Prevent locking of registration orders on payment checkout step
This is accomplished with a new checkout flow plugin and a new checkout flow using that plugin.

The only differences between this new plugin and MultistepDefault, which it is modeled on, are:

1. There is no 'review' step, since we weren't using it in the `registration` checkout flow.
2. CheckoutFlowBase::onStepChange() is overridden and the order lock/unlock code has been removed.

This does not seem to negatively effect the checkout process, and since the order remains unlocked when users enter the payment step, they can still go back to the site and view the cart.