More WIP: Switch to a form for email entry
The idea is for the user to end up in the cart (or even the checkout page). When the billing info is saved, some hook will check the associated registration for the following:

- There is only one participant
- That participant is missing required fields (like name)

If that's the case, the participant will be updated with the billing profile data and the checkout will continue as usual.

The other idea is that there will be a new route like /class/{salesforce_id}/single-registration that will redirect to this form. This will handle links from the marketing site.