Add a view for marketing to see orders with tracked utm codes
Since the data is stored in a 'map' field type, a simple new field formatter and an alteration of the filter were required.

The formatter just dumps the data as a YAML structure. YAML was chosen because it's easier to look at.

The filter was changed from 'standard' to 'string' to allow for crude filtering. It works for the purpose of this view, but probably wouldn't for other uses.