Prevent OrderVersionMismatchException when user logs in during checkout
Because the registration is saved during erf_user_login(), the order version was updated during checkout. We think that the order is also somehow stored or serialized as part of the form_state in the checkout flow, and this embedded order is then updated at the checkout step after login.

Co-authored-by: Aaron Froehlich <aaronfroehlich@gmail.com>