Move ILR outreach discount code temp storage to sessions
We had multiple issues when storing the applicable codes in the order data, specifically when adding an item while a new order/cart was also being created.

We also tried using tempstore.private, but that will not migrate a tempstore from an anonymous user when they log in. We may revisit that if https://drupal.org/i/3308135 is ever addressed.

While there are mentions that direct use of $_SESSION is discouraged, it is simple and works better than any of the other options.

Co-authored-by: Aaron Froehlich <aaronfroehlich@gmail.com>