Remove --partial usage and requirement to drush config import
This was done to allow the deletion of block config for the seven theme.

Originally, --partial was used and required to prevent the accidental removal of registration and participant types that were created on prod. During deployment, the config import would remove these config entities.

In practice, we never create participant or registration types on production.

Co-authored-by: Aaron Froehlich <aaronfroehlich@gmail.com>