Upgrade drupal/commerce to 2.39.0 with all dependencies
Commerce continues to be a PITA to upgrade properly. See https://docs.drupalcommerce.org/v2/installation/#updating

So this time, I just ran:

`composer require drupal/commerce:^2.39 --with-all-dependencies`

Which also upgrades:

- commerceguys/addressing (v1.4.2 => v2.2.1)
- drupal/token (1.13.0 => 1.14.0)
- drupal/state_machine (1.10.0 => 1.11.0)
- drupal/profile (1.7.0 => 1.10.0)
- drupal/inline_entity_form (1.0.0-rc15 => 1.0.0-rc17)
- drupal/entity_reference_revisions (1.10.0 => 1.11.0)
- drupal/address (1.12.0 => 2.0.2)

This required the removal of a patch to the drupal/address module from commit 9c115d7. We may need to review the issue at https://www.drupal.org/project/address/issues/2852800 to see if we need to update the order display from the above commit.